+++ src/main/java/com/cjm721/overloaded/cb/block/CompressedBlockHandler.java
@@ -30,6 +30,7 @@ public class CompressedBlockHandler {
         if(entry.recipeEnabled) {
           CompressedBlockAssets.addToRecipes(unCompressed.getRegistryName(), compressedBlock.getRegistryName());
         }
+        CompressedBlockAssets.addToDropLootTable(compressedBlock.getRegistryName());
 
         compressedBlock.setUnCompressed(unCompressed);
         if(unCompressed instanceof BlockCompressed) {